add entity reader to simplify prefabs
